Add unpacking root tarball to btrfs-formatted zram as new loop type.
Zram offers significant space savings compared to tmpfs.  Compression on
the root tarballs I have tested with approaches 50%.

BTRFS as the filesystem is due to personal plans for being able to boot
a filesystem image, configure physical disks, etc, and then migrate
storage from memory to permanent media with no reboot required.
Selection of zram filesystem is likely a desireable choice for future.

I expect this additionally needs the ability to verify the presence of
zram kernel feature, possibly the ability to detect lack of zram device
and dynamically allocate one, preferably the ability to dynamically
include mkfs.btrfs only when btrfs is the selected FS type.  Determining
the best way to add these features will be the subject of future
commits.